Added MobileComplete component
